Move data collection functions to their own module
This refactoring is useful to reduce the complexity
of the IonCore.js module. It also paves the way for
the additional collection features that will land
in future commits.
